Add C parser and generator
