+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
+## [0.0.2]
+
+- Add C parser and generator
+
 ## [0.0.1]
+
 - Initial release
